first argument of _BitScanReverse in MSVC is unsigned long ptr

+++ XbSymbolDatabase.c
@@ -41,7 +41,7 @@
 #include <intrin.h>
 static inline uint32_t BitScanReverse(uint32_t value)
 {
-    uint32_t index;
+    unsigned long index;
     if (!_BitScanReverse(&index, value)) {
         return 32;
     }
